* better-deploy:
  Differentiate betwen implicitly and explicitly chosen branches for deployments
  Deploy current branch unless specified
  Parameterize deploy.sh
  Refactor
  Prune docker entities after startup

+++ bin/deploy.sh
@@ -8,39 +8,46 @@ IFS=$'\n\t'
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 cd "$DIR"/..
 
+CONTAINER="$PROJECT_NAME"
+PORT="$INTERNAL_PORT"
+NETWORK="$CONTAINER"_net
+DEPLOY_BRANCH="${1:-}"
+BRANCH="$(git rev-parse --abbrev-ref HEAD)"
 set +x  # Do not print contents of .env
 source .env
 set -x
 
-if [ "$ENV" = "production" ]; then
+if [ -n "$DEPLOY_BRANCH" ]; then
     # Update repository
-    git checkout master
+    git checkout "$DEPLOY_BRANCH"
     git fetch -tp
     git pull
 fi
 
-# Build and start container
+# Build container and network
 docker pull "$(grep FROM Dockerfile | awk '{print $2}')"
-docker build -t "$PROJECT_NAME:$ENV" .
-docker network inspect "$PROJECT_NAME" &>/dev/null ||
-    docker network create --driver bridge "$PROJECT_NAME"
-docker stop "$PROJECT_NAME" || true
-docker container prune --force --filter "until=168h"
-docker image prune --force --filter "until=168h"
-docker volume prune --force
-docker container rm "$PROJECT_NAME" || true
+docker build -t "$CONTAINER:$BRANCH" .
+docker network inspect "$NETWORK" &>/dev/null ||
+    docker network create --driver bridge "$NETWORK"
+
+# Start container
+docker stop "$CONTAINER" || true
+docker container rm "$CONTAINER" || true
 docker run \
     --detach \
     --restart=always \
-    --publish="127.0.0.1:$INTERNAL_PORT:$INTERNAL_PORT" \
-    --network="$PROJECT_NAME" \
+    --publish="127.0.0.1:$PORT:$PORT" \
+    --network="$NETWORK" \
     --mount type=bind,source="$(pwd)"/app/static,target=/var/www/app/app/static \
     --mount type=bind,source="$(pwd)"/logs,target=/var/www/app/logs \
-    --name="$PROJECT_NAME" "$PROJECT_NAME:$ENV"
+    --name="$CONTAINER" "$CONTAINER:$BRANCH"
 
-if [ "$ENV" = "production" ]; then
+if [ "$ENV" = "production" ] && [ "$BRANCH" = "master" ]; then
     # Cleanup docker
+    docker container prune --force --filter "until=168h"
     docker image prune --force --filter "until=168h"
+    docker volume prune --force
+    docker network prune --force
 
     # Update nginx
     sudo service nginx reload
